% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_vowels.R
\name{mat2tbl}
\alias{mat2tbl}
\title{Convert a Vowel Pattern Matrix to a Tibble}
\usage{
mat2tbl(in_mat)
}
\arguments{
\item{in_mat}{A numeric matrix with rownames corresponding to vowel labels
and columns \code{F1_shift} and \code{F2_shift}.}
}
\value{
A tibble with columns \code{vowel}, \code{F1_shift}, \code{F2_shift}.
}
\description{
It is convenient to switch between tibble representations and
n x 2 matrix representations of covariation patterns. This
function does the convension from matric to tibble.
}
\examples{
# Generate a pattern with only three non-zero loadings for
# given list of vowels.
vowels <- c("BED", "BATH", "BEYOND")
p_1 <- generate_orthogonal_pattern(vowels, sparse = 3)
# Generate another orthogonal to the first with four non-trivial
# entries.
mat2tbl(p_1)
}
