% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.vowels-package.R
\docType{package}
\name{nzilbb.vowels-package}
\alias{nzilbb.vowels}
\alias{nzilbb.vowels-package}
\title{nzilbb.vowels: Vowel Covariation Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to support research on vowel covariation. Methods are provided to support Principal Component Analysis workflows (as in Brand et al. (2021) \doi{10.1016/j.wocn.2021.101096} and Wilson Black et al. (2023) \doi{10.1515/lingvan-2022-0086}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nzilbb.github.io/nzilbb_vowels/}
  \item \url{https://github.com/nzilbb/nzilbb_vowels}
  \item Report bugs at \url{https://github.com/nzilbb/nzilbb_vowels/issues}
}

}
\author{
\strong{Maintainer}: Joshua Wilson Black \email{joshua.black@canterbury.ac.nz} (\href{https://orcid.org/0000-0002-8272-5763}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Joshua Wilson Black \email{joshua.black@canterbury.ac.nz} (\href{https://orcid.org/0000-0002-8272-5763}{ORCID}) [copyright holder]
  \item James Brand \email{james.brand.ac@gmail.com} (\href{https://orcid.org/0000-0002-2853-9169}{ORCID})
}

}
\keyword{internal}
