% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onze_vowels}
\alias{onze_vowels}
\title{Monophthong data for random sample of speakers from the ONZE corpus.}
\format{
A data frame with 101572 rows and 8 variables:
\describe{
  \item{speaker}{Anonymised speaker code.}
  \item{vowel}{Factor variable with Wells lexical sets for 10 NZE monophthongs. Levels:  DRESS, FLEECE, GOOSE, KIT, LOT, NURSE, START, STRUT, THOUGHT, TRAP.}
  \item{F1_50}{First formant, extracted from vowel mid-point using LaBB-CAT interface with Praat.}
  \item{F2_50}{Second formant, extracted from vowel mid-point using LaBB-CAT interface with Praat.}
  \item{speech_rate}{Average speaker speech rate for whole recording.}
  \item{gender}{Gender of speaker, two levels: "M", "F".}
  \item{yob}{Year of birth of speaker.}
  \item{word}{Anonymised word code.}
}
}
\source{
\url{https://osf.io/q4j29/}
}
\usage{
onze_vowels
}
\description{
A dataset containing the the first and second formants, speech rate,
gender, and year of birth for 100 random speakers from the ONZE corpus. Data
is present for the following NZE monophthongs, represented by Wells lexical
sets: DRESS, FLEECE, GOOSE, KIT, LOT, NURSE, START, STRUT, THOUGHT, TRAP. Data
for FOOT is excluded (as in Brand et al. 2021).
}
\details{
Dataset is derived from the data made available in the supplementary materials
for Brand et al. (2021).
}
\keyword{datasets}
