% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_contrib_plot.R
\name{pca_contrib_plot}
\alias{pca_contrib_plot}
\title{Create plot of variables by contribution to given PC.}
\usage{
pca_contrib_plot(pca_object, pc_no = 1, cutoff = 50)
}
\arguments{
\item{pca_object}{a pca object generated by `prcomp` or `princomp`.}

\item{pc_no}{the PC to be visualised. Default value is 1.}

\item{cutoff}{the cutoff value for interpretation of the PC. The default
value, 50, means we ignore contributions until they cumulatively account
for more than 50% of the PC.}
}
\value{
`ggplot` object.
}
\description{
Plots match those given in supplementary material for Brand et al. (2021).
Variables are arranged by ascending contribution to the given PC, where
contribution is the squared loading for the variable expressed as a
percentage.
}
\examples{
pca_contrib_plot(pca_object, pc_no=1, cutoff=50)
pca_contrib_plot(pca_object, pc_no=2, cutoff=70)
}
