% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_test_plots.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Create plot of index loadings for a given PC from pca_test object.}
\usage{
plot_loadings(pca_test, pc_no = 1, violin = FALSE, filter_boots = FALSE)
}
\arguments{
\item{pca_test}{an object of class pca_test_results generated by `pca_test`.}

\item{pc_no}{An integer indicating which PC to plot.}

\item{violin}{If TRUE, violin plots are added for the confidence intervals of
the sampling distribution.}

\item{filter_boots}{if TRUE, only bootstrap iterations in which the variable
with the highest median loading is in the top 3/4 of its range are plotted.}
}
\value{
`ggplot` object.
}
\description{
Index loadings are presented with confidence intervals on the sampling
distribution generated by bootstrapping and a null distribution generated by
permutation.
}
\details{
If PCs are unstable, there is an option to take only the bootstrap iterations
in which the variable with the highest median loading across all iterations
is in the top 3/4 of its distribution. This helps to reveal
reliable connections of this variable with other variables in the data set.
}
\examples{
\dontrun{plot_loadings(pca_object, pc_no=1, cutoff=50)}
\dontrun{plot_loadings(pca_object, pc_no=2, cutoff=70)}
}
