% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qb_intervals}
\alias{qb_intervals}
\title{Mean formant, articulation rate, and amplitude values for intervals generated
for QuakeBox monologues. The same 77 speakers as in `qb_vowels` are used.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{interval_length}{Length of interval in seconds.}
  \item{speaker}{Anonymised speaker code.}
  \item{interval}{Time in seconds at which interval ends.}
  \item{articulation_rate}{Mean articulation rate within interval.}
  \item{amplitude}{Mean maximum amplitude within interval.}
  \item{F1_DRESS}{Speaker intercept from GAMM model of DRESS F1.}
  \item{F2_DRESS}{Speaker intercept from GAMM model of DRESS F2.}
  \item{F1_FLEECE}{Speaker intercept from GAMM model of FLEECE F1.}
  \item{F2_FLEECE}{Speaker intercept from GAMM model of FLEECE F2.}
  \item{F1_GOOSE}{Speaker intercept from GAMM model of GOOSE F1.}
  \item{F2_GOOSE}{Speaker intercept from GAMM model of GOOSE F2.}
  \item{F1_KIT}{Speaker intercept from GAMM model of KIT F1.}
  \item{F2_KIT}{Speaker intercept from GAMM model of KIT F2.}
  \item{F1_LOT}{Speaker intercept from GAMM model of LOT F1.}
  \item{F2_LOT}{Speaker intercept from GAMM model of LOT F2.}
  \item{F1_NURSE}{Speaker intercept from GAMM model of NURSE F1.}
  \item{F2_NURSE}{Speaker intercept from GAMM model of NURSE F2.}
  \item{F1_START}{Speaker intercept from GAMM model of START F1.}
  \item{F2_START}{Speaker intercept from GAMM model of START F2.}
  \item{F1_STRUT}{Speaker intercept from GAMM model of STRUT F1.}
  \item{F2_STRUT}{Speaker intercept from GAMM model of STRUT F2.}
  \item{F1_THOUGHT}{Speaker intercept from GAMM model of THOUGHT F1.}
  \item{F2_THOUGHT}{Speaker intercept from GAMM model of THOUGHT F2.}
  \item{F1_TRAP}{Speaker intercept from GAMM model of TRAP F1.}
  \item{F2_TRAP}{Speaker intercept from GAMM model of TRAP F2.}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
qb_intervals
}
\description{
QuakeBox monologues are divided into intervals of fixed length within mean
values are calcualted for formants, amplitude, and articulation rate.
}
\details{
Two interval lengths are given: 60 seconds and 240 seconds.

Formant data is z-scored by speaker and vowel, while the amplitude and
articulation rate are z-scored by speaker.

Original data was generated for Wilson Black et al. 2022.
}
\keyword{datasets}
