% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qb_vowels}
\alias{qb_vowels}
\title{Vowel data for 77 speakers from the QuakeBox corpus, with 11 in each age category.}
\format{
A data frame with 26331 rows and 14 variables:
\describe{
  \item{speaker}{Anonymised speaker code.}
  \item{vowel}{Factor variable with Wells lexical sets for 10 NZE monophthongs. Levels:  DRESS, FLEECE, GOOSE, KIT, LOT, NURSE, START, STRUT, THOUGHT, TRAP, FOOT.}
  \item{F1_50}{First formant in Hz, extracted from vowel mid-point using LaBB-CAT interface with Praat.}
  \item{F2_50}{Second formant in Hz, extracted from vowel mid-point using LaBB-CAT interface with Praat.}
  \item{participant_age_category}{Age category of speaker. Levels: 18-25, 26-35, 36-45, ..., 76-85.}
  \item{participant_gender}{Gender of participant. Levels: M, F.}
  \item{participant_nz_ethnic}{New Zealand ethnic category of participant. Levels: NZ mixed ethnicity, NZ European, Other.}
  \item{word_freq}{Frequency of word from which vowel token is taken in CELEX.}
  \item{word}{Anonymised word id. Allows for fitting models with word as random effect without revealing content of transcript.}
  \item{time}{Time in seconds at which vowel segment starts.}
  \item{vowel_duration}{Length of vowel in seconds.}
  \item{articulation_rate}{Articulation rate of utterance from which token is taken.}
  \item{following_segment_category}{Category of following segment. NB: liquids have already been removed. Levels: labial, velar, other.}
  \item{amplitude}{Maximum amplitude of word from which vowel token is taken, generated by LaBB-CAT interface with Praat.}
}
}
\source{
\url{https://github.com/JoshuaWilsonBlack/Amplitude_F1/}
}
\usage{
qb_vowels
}
\description{
A dataset containing formant values, amplitude, articulation rate, and
following segment data for 10 New Zealand English monophthongs, along with
participant demographics.
}
\details{
Original data was generated for Wilson Black et al. 2022.
}
\keyword{datasets}
