% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_vowels.R
\name{simulate_speakers}
\alias{simulate_speakers}
\title{Simulate Mid-Point Vocalic Data}
\usage{
simulate_speakers(
  grouped_means,
  n_speakers = 1000,
  probs = NULL,
  n_tokens_per_vowel = 30,
  min_tokens_per_vowel = NULL,
  patterns = list(),
  pattern_sds = numeric(0),
  speaker_noise_factor_sd = 0.25,
  scaling_factor_sd = 0.1
)
}
\arguments{
\item{grouped_means}{A tibble of grouped summary information about each vowel
category and grouping variable. The tibble must be structured as follows:
\itemize{
\item \code{vowel}: character vector identifying vowel types.
\item grouping variables: Optional character vectors identifying grouping
structure required for simulation (such as gender or age categories)
\item \code{mean_F1}: mean F1 value in Hz for group of speakers.
\item \code{sd_F1}: sd of F1 values in Hz for group of speakers.
\item \code{mean_F2}: mean F2 value in Hz for group of speakers.
\item \code{sd_F2}: sd of F2 values in Hz for group of speakers.
\item \code{cov_mat}: list column of (2x2) covariance matrices for within-vowel F1 and
F2 covariation.
}

Appropriate matrices can be generated by \code{\link[=generate_grouped_means]{generate_grouped_means()}}.}

\item{n_speakers}{Number of speakers to simulate. Default 1000.}

\item{probs}{List of probabilities passed to \code{sample()} to generate each
grouping factor. If none given, then each level is given equal
probabilities.}

\item{n_tokens_per_vowel}{Number of tokens to simulate per vowel. Default 30.
If NULL, use mean token count for vowel category from actual data with
\code{min_tokens_per_vowel} as control to ensure some tokens are present.}

\item{min_tokens_per_vowel}{Minimum tokens per vowel. Set to integer if
\code{n_tokens_per_vowel} is \code{NULL} to ensure a minimum number of tokens for your
analysis. Default NULL.}

\item{patterns}{List of pattern matrices (vowel x 2) representing covariation
patterns. See documentation for \code{\link[=generate_orthogonal_pattern]{generate_orthogonal_pattern()}} for the
required structure.}

\item{pattern_sds}{Numeric vector of standard deviations for each pattern
score.}

\item{speaker_noise_factor_sd}{SD of speaker-level noise factor. Default
0.25.}

\item{scaling_factor_sd}{SD of speaker-level formant scaling factor. Default
0.1.}
}
\value{
A tibble of simulated token-level formant data with speaker metadata.
}
\description{
Simulate F1 and F2 midpoint vocalic data for \code{n_speakers} speakers given a
collection of group means and standard deviations for any grouping structure
to be included in the simulation. Within-vowel covariation between F1 and F2
is automatically included. Any other patterns of covariation between
variables can be included by means of the \code{patterns} argument. The scale of
variation between speakers is controlled by the three standard deviation
arguments, which control variation due to induced covariation patterns,
random noise applied to each vowel category for each speaker, and how much
variance in vocal tract length there is, represented by a single scaling
parameter.
}
\examples{
# Generate grouped means by gender from real corpus.
g_m <- generate_grouped_means(
 qb_vowels,
 grouping_variables = "participant_gender"
)
# Simulate 4 speakers, with 10 vowels per speaker and no between-vowel
# covariation
simulate_speakers(
  g_m,
  n_speakers = 4,
  n_tokens_per_vowel = 10
)
}
