% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_vowels.R
\name{simulate_vs}
\alias{simulate_vs}
\title{Simulate vowel spaces from speaker means}
\usage{
simulate_vs(speaker_means)
}
\arguments{
\item{speaker_means}{A data frame containing columns \code{cov_mat},
\code{mean_F1}, and \code{mean_F2}.}
}
\value{
A data frame with columns \code{sim_f1} and \code{sim_f2} added,
unnested to one row per simulated observation.
}
\description{
Simulate vowel spaces from speaker means
}
