% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_vowels.R
\name{tbl2mat}
\alias{tbl2mat}
\title{Convert a Pattern Tibble to a Matrix}
\usage{
tbl2mat(in_tbl)
}
\arguments{
\item{in_tbl}{A tibble with columns \code{vowel}, \code{F1_shift}, \code{F2_shift}.}
}
\value{
A numeric matrix with rownames set to vowel labels.
}
\description{
It is convenient to switch between tibble representations and
n x 2 matrix representations of covariation patterns. This
function does the convension from tibble to matrix.
}
\examples{
# Generate a pattern with only three non-zero loadings for
# given list of vowels.
vowels <- c("BED", "BATH", "BEYOND")
p_1 <- generate_orthogonal_pattern(vowels, sparse = 3)
# Generate another orthogonal to the first with four non-trivial
# entries.
tbl_pat <- mat2tbl(p_1)
tbl2mat(tbl_pat)
}
