% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobanov_2.R
\name{lobanov_2}
\alias{lobanov_2}
\title{Apply Lobanov 2.0 normalisation to first and second formant data.}
\usage{
lobanov_2(vowel_data)
}
\arguments{
\item{vowel_data}{a data frame with speaker, vowel, F1, and F2 columns.}
}
\value{
input dataframe with additional columns `F1_lob2` and `F2_lob2`,
  containing the lobanov normalised F1 and F2 values respectively.
}
\description{
We assume the same structure as the normalisation functions in  the `vowels`
package. The first four columns should be:
}
\details{
1. speaker identifiers,
2. vowel identifiers,
3. first formant values in Hertz,
4. second formant values in Hertz.
}
\examples{
lobanov_2(onze_vowels)
}
