% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.vowels-package.R
\docType{package}
\name{nzilbb.vowels-package}
\alias{nzilbb.vowels}
\alias{nzilbb.vowels-package}
\title{nzilbb.vowels: Functions and Data for Vowel Covariation Studies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of tools to support research on vocalic covariation, especially research using Principal Component Analysis. Functions are provided to test whether PCA is appropriate, to determine how many PCs to report, and to choose which loadings to focus on in interpretation. Plotting functions are also provided to display PCA results in the vowel space. Datasets are also provided from recent work at the New Zealand Institute of Language, Brain and Behaviour.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nzilbb.github.io/nzilbb_vowels/}
  \item \url{https://github.com/nzilbb/nzilbb_vowels}
  \item Report bugs at \url{https://github.com/nzilbb/nzilbb_vowels/issues/}
}

}
\author{
\strong{Maintainer}: Joshua Wilson Black \email{joshua@wilsonblack.nz} (\href{https://orcid.org/0000-0002-8272-5763}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item James Brand \email{james.brand.ac@gmail.com} (\href{https://orcid.org/0000-0002-2853-9169}{ORCID})
}

}
\keyword{internal}
