% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.vowels-package.R
\docType{package}
\name{nzilbb.vowels-package}
\alias{nzilbb.vowels}
\alias{nzilbb.vowels-package}
\title{nzilbb.vowels: Functions and Data For Vowel Covariation Studies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data and functions to illustrate methods developed in recent studies Vowel covariation at the New Zealand Institute of Language, Brain, and Behaviour.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://joshuawilsonblack.github.io/nzilbb_vowels/}
}

}
\author{
\strong{Maintainer}: Joshua Wilson Black \email{joshua@wilsonblack.nz} (\href{https://orcid.org/0000-0002-8272-5763}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item James Brand \email{james.brand@ff.cuni.cz} (\href{https://orcid.org/0000-0002-2853-9169}{ORCID})
}

}
\keyword{internal}
