% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_contrib_plot.R
\name{pca_contrib_plot}
\alias{pca_contrib_plot}
\title{Create plot of variables by contribution to given PC.}
\usage{
pca_contrib_plot(pca_object, pc_no = 1, cutoff = 50)
}
\arguments{
\item{pca_object}{a pca object generated by \code{prcomp} or \code{princomp}.}

\item{pc_no}{the PC to be visualised. Default value is 1.}

\item{cutoff}{the cutoff value for interpretation of the PC. Determines what
total percentage contribution we want from the variables we select for
interpretation. The default of 50 means that we pick the variables with the
highest contribution to the PC until we have accounted for 50\% of the total
contributions to the PC. Can be set to \code{NULL} in which case, no cutoff value
is plotted.}
}
\value{
\code{ggplot} object.
}
\description{
Plots match those given in supplementary material for Brand et al. (2021).
Variables are arranged by ascending contribution to the given PC, where
contribution is the squared loading for the variable expressed as a
percentage.
}
\examples{
\dontrun{pca_contrib_plot(pca_object, pc_no=1, cutoff=50)}
\dontrun{pca_contrib_plot(pca_object, pc_no=2, cutoff=70)}
}
