% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plots.R
\name{plot_correlation_counts}
\alias{plot_correlation_counts}
\title{Create plot of correlation counts from `correlation_test` object.}
\usage{
plot_correlation_counts(cor_test, alpha = 0.05, half_violin = FALSE)
}
\arguments{
\item{cor_test}{an object of class `correlation_test` generated by
`correlation_test`.}

\item{alpha}{significance level for counting correlation as significant.}

\item{half_violin}{Plot correlation counts using a half violin plot and half
point plot.}
}
\value{
`ggplot` object.
}
\description{
This plot presents the distribution of counts of significant correlations at
a given alpha level and the count of significant correlations in the
original data.
}
\examples{
\dontrun{plot_correlation_counts(cor_test)}
}
