% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plots.R
\name{plot_correlation_magnitudes}
\alias{plot_correlation_magnitudes}
\title{Create plot of correlation magnitude from \code{correlation_test} object.}
\usage{
plot_correlation_magnitudes(cor_test)
}
\arguments{
\item{cor_test}{an object of class \code{correlation_test} generated by
\code{correlation_test}.}
}
\value{
\code{ggplot} object.
}
\description{
This plot type is used in
\insertCite{brandSystematicCovariationMonophthongs2021;textual}{nzilbb.vowels}.
It presents the magnitudes of the correlations from the real data as a solid
red line, and the correlations from each iteration of the permutation test as
light blue lines. This gives a visual sense of the distribution of random
correlations compared with those in the actual data.
}
\examples{
\dontrun{plot_correlation_magnitudes(cor_test)}
}
\references{
\insertAllCited{}
}
