% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_permutation_test.R
\name{plot_permutation_test}
\alias{plot_permutation_test}
\title{Create plot of results of permutation test.}
\usage{
plot_permutation_test(permutation_results)
}
\arguments{
\item{permutation_results}{object of class `permutation_results`.}
}
\value{
`ggplot` object.
}
\description{
Plots results of a permutation test carried out with the `permutation_test`
function.
}
\examples{
plot_permutation_test(permutation_results)
}
