% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qb_vowels}
\alias{qb_vowels}
\title{Vowel data for 77 speakers from the QuakeBox corpus, with 11 in each age category.}
\format{

}
\source{
\url{https://github.com/JoshuaWilsonBlack/Amplitude_F1/}
}
\usage{
qb_vowels
}
\description{
A dataset containing formant values, amplitude, articulation rate, and
following segment data for 10 New Zealand English monophthongs, along with
participant demographics.
}
\details{
Original data was generated for Wilson Black et al. 2022.
}
\keyword{datasets}
